Resolved merge conflict
